pyproject.toml
[tool.poetry]
name = "pearl"
version = "0.1.0"
description = ""
authors = ["hiranohachiman <shinhirano@keio.jp>"]
readme = "README.md"
packages = [{include = "pearl"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "2.2.0"
torchaudio = "2.2.0"
torchvision = "0.17.0"
torchmetrics = "0.2.0"
click = "^8.0"
pyyaml = "^6.0"
pytorch-lightning = "1.2.10"
pytorch-nlp = "^0.5.0"
pandas = "^1.5"
numpy = "^1.23"
scipy = "^1.9"
tqdm = "^4.64"
fairseq = "^0.12.2"
smart-open = "^6.0"
Pillow = "^9.0"
sentencepiece = "^0.1.99"
tensorboard = "^2.10"
wandb = "^0.15"
matplotlib = "^3.5"
termcolor = "^2.0"
ftfy = "^6.0"
pycocoevalcap = "^1.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
